Convert mins to f64
Avoids needing to convert data types, which would probably be slower than passing around an f64 anyway.

+++ src/structs.rs
@@ -6,8 +6,8 @@ pub struct PairData {
     pub ask_volume: f64,
     pub price_decimals: u8,
     pub volume_decimals: u8,
-    pub order_min: f32,
-    pub cost_min: f32,
+    pub order_min: f64,
+    pub cost_min: f64,
     pub pair_status: bool,
 }
 pub type PairDataVec = Vec<PairData>;
